Add versions of the Linq 'OpOrDefault' operators which return Maybe.
A number of Linq operators defined in System.Linq.Enumerable have an
associated operator which returns the default for the input type in
the case of failure e.g. First/FirstOrDefault.

Since default(T) is a valid value for T when using structs, these cannot
be used uniformly with value and reference types. Returning a Maybe<T>
allows failure to be represented in the same way for all types.